pyproject.toml
[tool.poetry]
name = "labor-costs"
version = "0.1.0"
description = ""
authors = ["pantyukhinda <dmitriy.pantyukhin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.119.1"
alembic = "^1.17.0"
sqlalchemy = "^2.0.44"
pydantic-core = "^2.41.4"
uvicorn = "^0.38.0"
pydantic-settings = "^2.11.0"
asyncpg = "^0.30.0"
passlib = "^1.7.4"
python-jose = "^3.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
